reduce amount of warnings

+++ src/osgEarth/ElevationLayer.cpp
@@ -537,7 +537,7 @@ ElevationLayerVector::createHeightField(const TileKey&                  key,
     //If any of the layers produced valid data then it's not considered a fallback
     if ( out_isFallback )
     {
-        *out_isFallback = (numFallbacks == heightFields.size());
+        *out_isFallback = (numFallbacks == (int)heightFields.size());
         //OE_NOTICE << "Num fallbacks=" << numFallbacks << " numHeightFields=" << heightFields.size() << " is fallback " << *out_isFallback << std::endl;
     }
